backend/requirements.txt
fastapi
uvicorn[standard]
openai
chromadb
python-dotenv
langgraph
langchain-openai
